Refactor: moving handleValidationResult to its own internal utility class.
